feat(split-mmap): chunk handling witht_ memory
